+++ selene-lib/tests/lints/roblox_incorrect_roact_usage/mixed_roact_react_usage.lua
@@ -1,0 +1,18 @@
+local e = React.createElement
+local x = Roact.createElement
+
+e("Frame", {
+    key = "",
+    ref = a,
+    children = {},
+    Name = "",
+    ThisPropertyDoesntExist = true,
+})
+
+x("Frame", {
+    key = "",
+    ref = a,
+    children = {},
+    Name = "",
+    ThisPropertyDoesntExist = true,
+})